move solution, recipient and price to select folder
