+++ app/pages/sections/catalogue-solutions/select/price/manifest.json
