+++ app/pages/sections/catalogue-solutions/select/recipient/controller.test.js
@@ -1,6 +1,6 @@
 import { getData } from 'buying-catalogue-library';
-import { orderApiUrl, solutionsApiUrl } from '../../../../config';
-import { logger } from '../../../../logger';
+import { orderApiUrl, solutionsApiUrl } from '../../../../../config';
+import { logger } from '../../../../../logger';
 import {
   getRecipientPageContext,
   getRecipients,
