+++ app/pages/sections/catalogue-solutions/select/solution/controller.js
@@ -1,6 +1,6 @@
 import { getData } from 'buying-catalogue-library';
-import { getEndpoint } from '../../../../endpoints';
-import { logger } from '../../../../logger';
+import { getEndpoint } from '../../../../../endpoints';
+import { logger } from '../../../../../logger';
 import { getContext, getErrorContext } from './contextCreator';
 
 export const getSolutionsPageContext = params => getContext(params);
