+++ app/pages/sections/catalogue-solutions/select/solution/controller.test.js
@@ -1,57 +1,34 @@
 import { getData } from 'buying-catalogue-library';
-import { solutionsApiUrl } from '../../../../config';
-import { logger } from '../../../../logger';
-import * as contextCreator from './contextCreator';
+import { solutionsApiUrl, orderApiUrl } from '../../../../../config';
+import { logger } from '../../../../../logger';
 import {
-  findSolutionPrices,
-  getSolutionPricePageContext,
-  validateSolutionPriceForm,
+  getSolutionsPageContext,
+  findSolutions,
+  getSupplierId,
+  validateSolutionForm,
 } from './controller';
+import * as contextCreator from './contextCreator';
 
 jest.mock('buying-catalogue-library');
-jest.mock('../../../../logger');
+
 jest.mock('./contextCreator', () => ({
   getContext: jest.fn(),
 }));
 
-const accessToken = 'access_token';
-const orderId = 'order-id';
-const solutionId = 'sol-1';
-
-const solutionPrices = {
-  id: 'sol-1',
-  name: 'name',
-  prices: [
-    {
-      type: 'flat',
-      currencyCode: 'GBP',
-      itemUnit: {
-        name: 'patient',
-        description: 'per patient',
-      },
-      timeUnit: {
-        name: 'year',
-        description: 'per year',
-      },
-      price: 1.64,
-    },
-  ],
-};
-
-describe('select-price controller', () => {
-  describe('getSolutionPricePageContext', () => {
+describe('catalogue-solutions select-solution controller', () => {
+  describe('getSolutionsPageContext', () => {
     it('should call getContext with the correct params', async () => {
       contextCreator.getContext
         .mockResolvedValueOnce();
 
-      await getSolutionPricePageContext({ orderId, solutionPrices });
+      await getSolutionsPageContext({ orderId: 'order-1' });
 
       expect(contextCreator.getContext.mock.calls.length).toEqual(1);
-      expect(contextCreator.getContext).toHaveBeenCalledWith({ orderId, solutionPrices });
+      expect(contextCreator.getContext).toHaveBeenCalledWith({ orderId: 'order-1' });
     });
   });
 
-  describe('findSolutionPrices', () => {
+  describe('findSolutions', () => {
     afterEach(() => {
       getData.mockReset();
     });
@@ -60,11 +37,33 @@ describe('select-price controller', () => {
       getData
         .mockResolvedValueOnce({ data: {} });
 
-      await findSolutionPrices({ accessToken, solutionId });
+      await findSolutions({ supplierId: 'supp-1', accessToken: 'access_token' });
       expect(getData.mock.calls.length).toEqual(1);
       expect(getData).toHaveBeenCalledWith({
-        endpoint: `${solutionsApiUrl}/api/v1/solutions/${solutionId}/pricing`,
+        endpoint: `${solutionsApiUrl}/api/v1/solutions?supplierId=supp-1`,
         accessToken: 'access_token',
+        logger,
+      });
+    });
+  });
+
+  describe('getSupplierId', () => {
+    afterEach(() => {
+      getData.mockReset();
+      contextCreator.getContext.mockReset();
+    });
+
+    const accessToken = 'access_token';
+    const orderId = 'order-id';
+
+    it('should call getData with the correct params when hasSavedData is true', async () => {
+      getData.mockResolvedValueOnce({ supplierId: 'supp-1' });
+
+      await getSupplierId({ orderId, accessToken });
+      expect(getData.mock.calls.length).toEqual(1);
+      expect(getData).toHaveBeenCalledWith({
+        endpoint: `${orderApiUrl}/api/v1/orders/order-id/sections/supplier`,
+        accessToken,
         logger,
       });
     });
@@ -74,10 +73,10 @@ describe('select-price controller', () => {
     describe('when there are no validation errors', () => {
       it('should return success as true', () => {
         const data = {
-          selectSolutionPrice: 'some-solution-id',
+          selectSolution: 'some-solution-id',
         };
 
-        const response = validateSolutionPriceForm({ data });
+        const response = validateSolutionForm({ data });
 
         expect(response.success).toEqual(true);
       });
@@ -86,8 +85,8 @@ describe('select-price controller', () => {
     describe('when there are validation errors', () => {
       const expectedValidationErrors = [
         {
-          field: 'selectSolutionPrice',
-          id: 'SelectSolutionPriceRequired',
+          field: 'selectSolution',
+          id: 'SelectSolutionRequired',
         },
       ];
 
@@ -96,7 +95,7 @@ describe('select-price controller', () => {
           selectSolution: '',
         };
 
-        const response = validateSolutionPriceForm({ data });
+        const response = validateSolutionForm({ data });
 
         expect(response.success).toEqual(false);
         expect(response.errors).toEqual(expectedValidationErrors);
@@ -107,7 +106,7 @@ describe('select-price controller', () => {
           selectSolution: '   ',
         };
 
-        const response = validateSolutionPriceForm({ data });
+        const response = validateSolutionForm({ data });
 
         expect(response.success).toEqual(false);
         expect(response.errors).toEqual(expectedValidationErrors);
@@ -116,7 +115,7 @@ describe('select-price controller', () => {
       it('should return a validation error if supplierName is undefined', () => {
         const data = {};
 
-        const response = validateSolutionPriceForm({ data });
+        const response = validateSolutionForm({ data });
 
         expect(response.errors).toEqual(expectedValidationErrors);
       });
