+++ app/pages/sections/catalogue-solutions/select/solution/manifest.json
