+++ app/pages/sections/catalogue-solutions/select/solution/ui-tests/general.ui.test.js
@@ -2,7 +2,7 @@ import nock from 'nock';
 import { ClientFunction, Selector } from 'testcafe';
 import { extractInnerText } from 'buying-catalogue-library';
 import content from '../manifest.json';
-import { solutionsApiUrl, orderApiUrl } from '../../../../../config';
+import { solutionsApiUrl, orderApiUrl } from '../../../../../../config';
 
 const pageUrl = 'http://localhost:1234/order/organisation/order-id/catalogue-solutions/solution';
